feat(framework,medusa): Ensure publishable key middleware is set for all store endpoints
